feat: extract static components from list client bundles
Following the pattern from franksmovielog.com PR #2297, this refactors
the list components to separate static layout from interactive elements.

Changes:
- Created simplified server-side ListWithFiltersLayout for static rendering
- Created new ListWithFilters component for all interactive functionality
- Updated list components (Reviews, Authors, Readings, Author) to remove layout
- Modified Astro pages to render static layout server-side
- Only hydrate interactive components client-side

This reduces client bundle sizes by extracting static markup (backdrop,
layout, footer) from JavaScript bundles, similar to the movielog
optimization that reduced page chunks from 95KB to 1-7KB.
