+++ src/components/Authors/Authors.tsx
@@ -2,13 +2,11 @@ import { type JSX, useReducer } from "react";
 
 import type { Author } from "~/api/authors";
 import type { AvatarImageProps } from "~/api/avatars";
-import type { BackdropImageProps } from "~/api/backdrops";
 
-import { Backdrop } from "~/components/Backdrop";
 import { GroupedList } from "~/components/GroupedList";
 import { ListItem } from "~/components/ListItem";
 import { ListItemAvatar } from "~/components/ListItemAvatar";
-import { ListWithFiltersLayout } from "~/components/ListWithFiltersLayout";
+import { ListWithFilters } from "~/components/ListWithFilters";
 
 import type { Sort } from "./Authors.reducer";
 
@@ -19,19 +17,17 @@ export type ListItemValue = Pick<Author, "name" | "slug" | "sortName"> & {
   reviewedWorkCount: number;
 };
 
-export type Props = {
-  backdropImageProps: BackdropImageProps;
-  deck: string;
+export type InteractiveProps = {
   initialSort: Sort;
   values: ListItemValue[];
 };
 
-export function Authors({
-  backdropImageProps,
-  deck,
-  initialSort,
-  values,
-}: Props): JSX.Element {
+export type Props = InteractiveProps & {
+  backdropImageProps: any;
+  deck: string;
+};
+
+export function Authors({ initialSort, values }: InteractiveProps): JSX.Element {
   const [state, dispatch] = useReducer(
     reducer,
     {
@@ -42,10 +38,7 @@ export function Authors({
   );
 
   return (
-    <ListWithFiltersLayout
-      backdrop={
-        <Backdrop deck={deck} imageProps={backdropImageProps} title="Authors" />
-      }
+    <ListWithFilters
       filters={<Filters dispatch={dispatch} sortValue={state.sortValue} />}
       list={
         <GroupedList
