+++ src/components/Readings/Readings.tsx
@@ -1,17 +1,17 @@
 import { type JSX, useReducer } from "react";
 
-import type { BackdropImageProps } from "~/api/backdrops";
 import type { CoverImageProps } from "~/api/covers";
 import type { TimelineEntry } from "~/api/timelineEntries";
 
-import { Backdrop } from "~/components/Backdrop";
 import { BarGradient } from "~/components/BarGradient";
 import { GroupedList } from "~/components/GroupedList";
 import { ListItem } from "~/components/ListItem";
 import { ListItemCover } from "~/components/ListItemCover";
 import { ListItemKindAndYear } from "~/components/ListItemKindAndYear";
-import { ListWithFiltersLayout } from "~/components/ListWithFiltersLayout";
-import { ListHeaderButton } from "~/components/ListWithFiltersLayout";
+import {
+  ListHeaderButton,
+  ListWithFilters,
+} from "~/components/ListWithFilters";
 import { toSentenceArray } from "~/utils";
 
 import type { Sort } from "./Readings.reducer";
@@ -37,11 +37,9 @@ export type ListItemValue = Pick<
   readingMonth: string;
   readingYear: string;
 };
-export type Props = {
+export type InteractiveProps = {
   abandonedCount: number;
-  backdropImageProps: BackdropImageProps;
   bookCount: number;
-  deck: string;
   distinctEditions: string[];
   distinctKinds: string[];
   distinctReadingYears: string[];
@@ -52,16 +50,19 @@ export type Props = {
   workCount: number;
 };
 
+export type Props = InteractiveProps & {
+  backdropImageProps: any;
+  deck: string;
+};
+
 export function Readings({
-  backdropImageProps,
-  deck,
   distinctEditions,
   distinctKinds,
   distinctReadingYears,
   distinctWorkYears,
   initialSort,
   values,
-}: Props): JSX.Element {
+}: InteractiveProps): JSX.Element {
   const [state, dispatch] = useReducer(
     reducer,
     {
@@ -72,14 +73,7 @@ export function Readings({
   );
 
   return (
-    <ListWithFiltersLayout
-      backdrop={
-        <Backdrop
-          deck={deck}
-          imageProps={backdropImageProps}
-          title="Reading Log"
-        />
-      }
+    <ListWithFilters
       filters={
         <Filters
           dispatch={dispatch}
