+++ src/components/Reviews/Reviews.tsx
@@ -1,18 +1,16 @@
 import { type JSX, useReducer } from "react";
 
-import type { BackdropImageProps } from "~/api/backdrops";
 import type { CoverImageProps } from "~/api/covers";
 import type { Review } from "~/api/reviews";
 
 import { Abandoned } from "~/components/Abandoned";
-import { Backdrop } from "~/components/Backdrop";
 import { Grade } from "~/components/Grade";
 import { GroupedList } from "~/components/GroupedList";
 import { ListItem } from "~/components/ListItem";
 import { ListItemCover } from "~/components/ListItemCover";
 import { ListItemKindAndYear } from "~/components/ListItemKindAndYear";
 import { ListItemTitle } from "~/components/ListItemTitle";
-import { ListWithFiltersLayout } from "~/components/ListWithFiltersLayout";
+import { ListWithFilters } from "~/components/ListWithFilters";
 import { toSentenceArray } from "~/utils";
 
 import type { Sort } from "./Reviews.reducer";
@@ -35,9 +33,7 @@ export type ListItemValue = Pick<
   coverImageProps: CoverImageProps;
 };
 
-export type Props = {
-  backdropImageProps: BackdropImageProps;
-  deck: string;
+export type InteractiveProps = {
   distinctKinds: readonly string[];
   distinctPublishedYears: readonly string[];
   distinctReviewYears: readonly string[];
@@ -45,17 +41,20 @@ export type Props = {
   values: ListItemValue[];
 };
 
+export type Props = InteractiveProps & {
+  backdropImageProps: any;
+  deck: string;
+};
+
 type Author = Pick<Review["authors"][0], "name" | "sortName"> & {};
 
 export function Reviews({
-  backdropImageProps,
-  deck,
   distinctKinds,
   distinctPublishedYears,
   distinctReviewYears,
   initialSort,
   values,
-}: Props): JSX.Element {
+}: InteractiveProps): JSX.Element {
   const [state, dispatch] = useReducer(
     reducer,
     {
@@ -66,10 +65,7 @@ export function Reviews({
   );
 
   return (
-    <ListWithFiltersLayout
-      backdrop={
-        <Backdrop deck={deck} imageProps={backdropImageProps} title="Reviews" />
-      }
+    <ListWithFilters
       filters={
         <Filters
           dispatch={dispatch}
