+++ src/pages/authors/[slug]/index.astro
@@ -1,7 +1,11 @@
 ---
+import React from "react";
+
 import { allAuthors } from "~/api/authors";
 import { Author } from "~/components/Author/Author";
 import { getProps } from "~/components/Author/getProps";
+import { AvatarBackdrop, BreadcrumbLink } from "~/components/Backdrop";
+import { ListWithFiltersLayout } from "~/components/ListWithFiltersLayout";
 import Layout from "~/layouts/Layout.astro";
 
 interface Props {
@@ -25,18 +29,43 @@ export async function getStaticPaths() {
 
 const { slug } = Astro.props;
 const props = await getProps(slug);
+
+// Separate layout props from interactive props
+const {
+  avatarImageProps,
+  backdropImageProps,
+  deck,
+  name,
+  ...interactiveProps
+} = props;
+
+// Create React elements in front matter
+const breadcrumbElement = React.createElement(BreadcrumbLink, {
+  children: "Authors",
+  href: "/authors/",
+});
+
+const backdropElement = React.createElement(AvatarBackdrop, {
+  avatarImageProps,
+  backdropImageProps,
+  breadcrumb: breadcrumbElement,
+  deck,
+  name,
+});
 ---
 
 <Layout
   canonical={true}
   meta={{
-    description: `Reviews of books and stories written by ${props.name}. Sort reviews by best or worst, newest or oldest. Filter by year, kind, or title.`,
+    description: `Reviews of books and stories written by ${name}. Sort reviews by best or worst, newest or oldest. Filter by year, kind, or title.`,
   }}
   openGraph={{
-    description: props.deck,
+    description: deck,
     type: "website",
   }}
-  title={`${props.name} book & story reviews`}
+  title={`${name} book & story reviews`}
 >
-  <Author client:load {...props} />
+  <ListWithFiltersLayout backdrop={backdropElement} hasBackdrop={true}>
+    <Author client:load {...interactiveProps} name={name} />
+  </ListWithFiltersLayout>
 </Layout>
