+++ src/pages/authors/index.astro
@@ -1,9 +1,23 @@
 ---
+import React from "react";
+
 import { Authors } from "~/components/Authors/Authors";
 import { getProps } from "~/components/Authors/getProps";
+import { Backdrop } from "~/components/Backdrop";
+import { ListWithFiltersLayout } from "~/components/ListWithFiltersLayout";
 import Layout from "~/layouts/Layout.astro";
 
 const props = await getProps();
+
+// Separate layout props from interactive props
+const { backdropImageProps, deck, ...interactiveProps } = props;
+
+// Create React elements in front matter
+const backdropElement = React.createElement(Backdrop, {
+  deck,
+  imageProps: backdropImageProps,
+  title: "Authors",
+});
 ---
 
 <Layout
@@ -13,7 +27,7 @@ const props = await getProps();
       "A sortable index of authors who contributed to the books and stories reviewed on this site. Sort by name or review count.",
   }}
   openGraph={{
-    description: props.deck,
+    description: deck,
     type: "website",
   }}
   title="Authors"
@@ -21,5 +35,7 @@ const props = await getProps();
   <Fragment slot="meta-theme-color">
     <meta content="rgba(0,0,0,.85)" name="theme-color" />
   </Fragment>
-  <Authors client:load {...props} />
+  <ListWithFiltersLayout backdrop={backdropElement}>
+    <Authors client:load {...interactiveProps} />
+  </ListWithFiltersLayout>
 </Layout>
